pyproject.toml
[tool.poetry]
name = "llamaindex-chat-with-streamlit-docs"
version = "0.1.0"
description = ""
authors = ["mattsalah <msalah@infraclear.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.41.1"
openai = "^1.59.7"
llama-index = "^0.12.10"
llama-index-llms-openai = "^0.3.13"
nltk = "^3.9.1"
pypdf = "^5.1.0"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"
loguru = "^0.7.3"
lxml = "^5.3.0"
boto3 = "^1.35.98"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
